chore: Refactor Solution Variable service.
